Matching pre-revisions
Prior to conversations with Scott on 11/7, the matching plan was an facsimile of the Alcohol Paper: Select variables (similar to alcohol), clean, carry-forward, split, pool, etc
We have since then pivoted, but should archive these changes in case we decide to go back.
Included a `_NOTES` to keep track of small changes and over-arching goal of analysis.